fix(backend): コードレビュー指摘修正（並行性・パフォーマンス・正しさ）
- windivert: Vec クローン除去・drain で二重アロケーション解消・spawn → await 直接呼び出し
- windivert: TCP シーケンス番号を u32 wrapping 演算に統一
- processor: BuffTick/BuffSnapshot 二重デコードを else-if に・Vec::remove(0) → スライス・attrs.clone() 除去
- capture/mod: is_paused の二重ロック解消
- history: lock poisoning で続行せず log::error + return に
- binary_reader: is_empty を remaining()==0 に修正
- name_cache: last_seen_ms 更新を常に保存対象に
- selected_uid: flush() でロック外にファイル書き込み
- lib: tray icon の unwrap() をパニック回避に
- commands: u128 → f64 キャストを saturating_sub で安全化

+++ src-tauri/src/capture/tcp_reassembler.rs
@@ -1,8 +1,8 @@
 use std::collections::BTreeMap;
 
 pub struct TcpReassembler {
-    pub cache: BTreeMap<usize, Vec<u8>>, // sequence -> payload
-    pub next_seq: Option<usize>,         // next expected sequence number
+    pub cache: BTreeMap<u32, Vec<u8>>, // sequence -> payload
+    pub next_seq: Option<u32>,         // next expected sequence number
     pub data: Vec<u8>,
 }
 
@@ -15,7 +15,7 @@ impl TcpReassembler {
         }
     }
 
-    pub fn clear(&mut self, seq_number: usize) {
+    pub fn clear(&mut self, seq_number: u32) {
         self.cache = BTreeMap::new();
         self.data.clear();
         self.next_seq = Some(seq_number);
